切換成讀取product Table物件

+++ app/src/main/java/com/example/alpha/ui/myObject/productItem.kt
@@ -4,4 +4,4 @@ data class ProductItem(
     val pId: Int,
     val imageUrl: Int, val pName: String,
     val pType: String,
-    val pNumber: Int, val pPrice: Double)
+    val pNumber: Int, val pPrice: Int)
